添加了删除用户的功能
挺鸡肋, 就是拿来证明我写了权限判断的

+++ User.cpp
@@ -83,7 +83,7 @@ namespace Hyt
 		AES aes(256);
 		unsigned char* c = aes.EncryptECB(plain, plainLen, key, outLen);
 
-		char* t = new char[outLen + 1];
+		char* t = new char[(unsigned long long)outLen + 1];
 		memcpy(t, c, outLen);
 		t[outLen] = 0;
